Refactor: 한 파일에 모여있던 코드를 큰 맥락의 관심사별로만 분리한다
