+++ src/utils/reactHookForm.ts
@@ -1,0 +1,9 @@
+import { ValueOf } from '@/types';
+import { FieldValues, UseFormTrigger } from 'react-hook-form';
+
+export const validateForm = <F extends FieldValues>(
+  trigger: UseFormTrigger<F>,
+  formKey: ValueOf<F>,
+) => {
+  trigger(formKey);
+};
